autogen-service/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
autogen-agentchat~=0.2
boto3>=1.34
requests>=2.32
pydantic==2.9.2
python-dotenv>=1.0
opentelemetry-distro>=0.48b0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
azure-cosmos>=4.5.1
